Added dropdowns for start time and end time

+++ FrontEnd/src/main/java/dao/EventDao.java
@@ -1,7 +1,6 @@
 package dao;
 
 import exception.DaoException;
-import model.Availability;
 import model.Event;
 
 import java.util.List;
@@ -15,5 +14,5 @@ public interface EventDao {
 
     List<Event> findEventbyId(int id);
 
-    void removeAndUpdateOptTime(int eventId, int optTime);
+    void removeAndUpdateOptTime(int eventId, double optTime);
 }
